+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added 
-- An optional `pageSize` to `/queue` command 
+- An optional `page-size` to `/queue` command 
 
 ## [2.9.3] - 2024-08-19